AHHA_240806.01

+++ AHHA.API/Controllers/Masters/CountryController.cs
@@ -1,13 +1,17 @@
-﻿using AHHA.Application.IServices.Masters;
+﻿using AHHA.Application.IServices;
+using AHHA.Application.IServices.Masters;
+using AHHA.Core.Common;
 using AHHA.Core.Entities.Masters;
-using AHHA.Core.Models;
+using AHHA.Core.Models.Masters;
 using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Caching.Memory;
 using Microsoft.Extensions.Primitives;
 
 namespace AHHA.API.Controllers.Masters
 {
+    [Authorize]
     [Route("api/[controller]")]
     [ApiController]
     public class CountryController : BaseController
@@ -18,9 +22,10 @@ namespace AHHA.API.Controllers.Masters
         private Int16 pageNumber = 1;
         private Int16 CompanyId = 0;
         private Int32 UserId = 0;
+        private Int32 RegId = 0;
 
-        public CountryController(IMemoryCache memoryCache, IMapper mapper, ILogger<CountryController> logger, ICountryService countryService)
-    : base(memoryCache, mapper)
+        public CountryController(IMemoryCache memoryCache, IMapper mapper, IBaseService baseServices, ILogger<CountryController> logger, ICountryService countryService)
+    : base(memoryCache, mapper, baseServices)
         {
             _logger = logger;
             _countryService = countryService;
@@ -33,28 +38,38 @@ namespace AHHA.API.Controllers.Masters
             {
                 CompanyId = Convert.ToInt16(Request.Headers.TryGetValue("CompanyId", out StringValues headerValue));
                 UserId = Convert.ToInt32(Request.Headers.TryGetValue("UserId", out StringValues userIdValue));
+                RegId = Convert.ToInt32(Request.Headers.TryGetValue("RegId", out StringValues regIdValue));
 
                 if (ValidateHeaders(CompanyId, UserId))
                 {
-                    pageSize = (Request.Headers.TryGetValue("pageSize", out StringValues pageSizeValue)) == true ? Convert.ToInt16(pageSizeValue[0]) : pageSize;
-                    pageNumber = (Request.Headers.TryGetValue("pageNumber", out StringValues pageNumberValue)) == true ? Convert.ToInt16(pageNumberValue[0]) : pageNumber;
+                    var userGroupRight = ValidateScreen(CompanyId, (Int16)Modules.Master, (Int32)Master.Country, UserId);
 
-                    //Get the data from cache memory
-                    var cacheData = _memoryCache.Get<CountryViewModelCount>("country");
+                    if (userGroupRight != null)
+                    {
+                        pageSize = (Request.Headers.TryGetValue("pageSize", out StringValues pageSizeValue)) == true ? Convert.ToInt16(pageSizeValue[0]) : pageSize;
+                        pageNumber = (Request.Headers.TryGetValue("pageNumber", out StringValues pageNumberValue)) == true ? Convert.ToInt16(pageNumberValue[0]) : pageNumber;
 
-                    if (cacheData != null)
-                        return Ok(cacheData);
+                        //Get the data from cache memory
+                        var cacheData = _memoryCache.Get<CountryViewModelCount>("country");
+
+                        if (cacheData != null)
+                            return Ok(cacheData);
+                        else
+                        {
+                            var expirationTime = DateTimeOffset.Now.AddMinutes(5);
+                            cacheData = await _countryService.GetCountryListAsync(CompanyId, pageSize, pageNumber, UserId);
+
+                            if (cacheData == null)
+                                return NotFound();
+
+                            _memoryCache.Set<CountryViewModelCount>("country", cacheData, expirationTime);
+
+                            return Ok(cacheData);
+                        }
+                    }
                     else
                     {
-                        var expirationTime = DateTimeOffset.Now.AddMinutes(5);
-                        cacheData = await _countryService.GetCountryListAsync(CompanyId, pageSize, pageNumber, UserId);
-
-                        if (cacheData == null)
-                            return NotFound();
-
-                        _memoryCache.Set<CountryViewModelCount>("country", cacheData, expirationTime);
-
-                        return Ok(cacheData);
+                        return NotFound("Users not have a access for this screen");
                     }
                 }
                 else
@@ -86,25 +101,37 @@ namespace AHHA.API.Controllers.Masters
 
                 if (ValidateHeaders(CompanyId, UserId))
                 {
+                    var userGroupRight = ValidateScreen(CompanyId, (Int16)Modules.Master, (Int32)Master.Country, UserId);
 
-                    if (_memoryCache.TryGetValue($"country_{CountryId}", out CountryViewModel? cachedProduct))
+                    if (userGroupRight != null)
                     {
-                        countryViewModel = cachedProduct;
+                        if (_memoryCache.TryGetValue($"country_{CountryId}", out CountryViewModel? cachedProduct))
+                        {
+                            countryViewModel = cachedProduct;
+                        }
+                        else
+                        {
+                            countryViewModel = _mapper.Map<CountryViewModel>(await _countryService.GetCountryByIdAsync(CompanyId, CountryId, UserId));
+
+                            if (countryViewModel == null)
+                                return NotFound();
+                            else
+                                // Cache the country with an expiration time of 10 minutes
+                                _memoryCache.Set($"Country_{CountryId}", countryViewModel, TimeSpan.FromMinutes(10));
+                        }
+
+                        return Ok(countryViewModel);
                     }
                     else
                     {
-                        countryViewModel = _mapper.Map<CountryViewModel>(await _countryService.GetCountryByIdAsync(CompanyId, CountryId, UserId));
-
-                        if (countryViewModel == null)
-                            return NotFound();
-                        else
-                            // Cache the country with an expiration time of 10 minutes
-                            _memoryCache.Set($"Country_{CountryId}", countryViewModel, TimeSpan.FromMinutes(10));
+                        return NotFound("Users not have a access for this screen");
                     }
-
-                    return Ok(countryViewModel);
                 }
-                else { return NoContent(); }
+                else
+                {
+                    return NoContent();
+                }
+
             }
             catch (Exception ex)
             {
@@ -124,22 +151,39 @@ namespace AHHA.API.Controllers.Masters
 
                 if (ValidateHeaders(CompanyId, UserId))
                 {
-                    if (country == null)
-                        return BadRequest();
+                    var userGroupRight = ValidateScreen(CompanyId, (Int16)Modules.Master, (Int32)Master.Country, UserId);
 
-                    var countryEntity = new M_Country
+                    if (userGroupRight != null)
                     {
-                        CompanyId = country.CompanyId,
-                        CountryCode = country.CountryCode,
-                        CountryId = country.CountryId,
-                        CountryName = country.CountryName,
-                        CreateById = UserId,
-                        IsActive = country.IsActive,
-                        Remarks = country.Remarks
-                    };
+                        if (userGroupRight.IsCreate)
+                        {
+                            if (country == null)
+                                return BadRequest();
 
-                    var createdCountry = await _countryService.AddCountryAsync(CompanyId, countryEntity, UserId);
-                    return Ok(createdCountry);
+                            var countryEntity = new M_Country
+                            {
+                                CompanyId = country.CompanyId,
+                                CountryCode = country.CountryCode,
+                                CountryId = country.CountryId,
+                                CountryName = country.CountryName,
+                                CreateById = UserId,
+                                IsActive = country.IsActive,
+                                Remarks = country.Remarks
+                            };
+
+                            var createdCountry = await _countryService.AddCountryAsync(CompanyId, countryEntity, UserId);
+                            return Ok(createdCountry);
+
+                        }
+                        else
+                        {
+                            return NotFound("Users do not have a access to delete");
+                        }
+                    }
+                    else
+                    {
+                        return NotFound("Users not have a access for this screen");
+                    }
                 }
                 else
                 {
@@ -165,36 +209,51 @@ namespace AHHA.API.Controllers.Masters
 
                 if (ValidateHeaders(CompanyId, UserId))
                 {
+                    var userGroupRight = ValidateScreen(CompanyId, (Int16)Modules.Master, (Int32)Master.Country, UserId);
 
-                    if (CountryId != country.CountryId)
-                        return BadRequest("M_Country ID mismatch");
-
-                    // Attempt to retrieve the country from the cache
-                    if (_memoryCache.TryGetValue($"country_{CountryId}", out CountryViewModel? cachedProduct))
+                    if (userGroupRight != null)
                     {
-                        countryViewModel = cachedProduct;
+                        if (userGroupRight.IsEdit)
+                        {
+                            if (CountryId != country.CountryId)
+                                return BadRequest("M_Country ID mismatch");
+
+                            // Attempt to retrieve the country from the cache
+                            if (_memoryCache.TryGetValue($"country_{CountryId}", out CountryViewModel? cachedProduct))
+                            {
+                                countryViewModel = cachedProduct;
+                            }
+                            else
+                            {
+                                var CountryToUpdate = await _countryService.GetCountryByIdAsync(CompanyId, CountryId, UserId);
+
+                                if (CountryToUpdate == null)
+                                    return NotFound($"M_Country with Id = {CountryId} not found");
+                            }
+
+                            var countryEntity = new M_Country
+                            {
+                                CountryCode = country.CountryCode,
+                                CountryId = country.CountryId,
+                                CountryName = country.CountryName,
+                                EditById = UserId,
+                                EditDate = DateTime.Now,
+                                IsActive = country.IsActive,
+                                Remarks = country.Remarks
+                            };
+
+                            var sqlResponce = await _countryService.UpdateCountryAsync(CompanyId, countryEntity, UserId);
+                            return Ok(sqlResponce);
+                        }
+                        else
+                        {
+                            return NotFound("Users do not have a access to delete");
+                        }
                     }
                     else
                     {
-                        var CountryToUpdate = await _countryService.GetCountryByIdAsync(CompanyId, CountryId, UserId);
-
-                        if (CountryToUpdate == null)
-                            return NotFound($"M_Country with Id = {CountryId} not found");
+                        return NotFound("Users not have a access for this screen");
                     }
-
-                    var countryEntity = new M_Country
-                    {
-                        CountryCode = country.CountryCode,
-                        CountryId = country.CountryId,
-                        CountryName = country.CountryName,
-                        EditById = UserId,
-                        EditDate = DateTime.Now,
-                        IsActive = country.IsActive,
-                        Remarks = country.Remarks
-                    };
-
-                    var sqlResponce = await _countryService.UpdateCountryAsync(CompanyId, countryEntity, UserId);
-                    return Ok(sqlResponce);
                 }
                 else
                 {
@@ -219,16 +278,31 @@ namespace AHHA.API.Controllers.Masters
 
                 if (ValidateHeaders(CompanyId, UserId))
                 {
+                    var userGroupRight = ValidateScreen(CompanyId, (Int16)Modules.Master, (Int32)Master.Country, UserId);
 
-                    var CountryToDelete = await _countryService.GetCountryByIdAsync(CompanyId, CountryId, UserId);
+                    if (userGroupRight != null)
+                    {
+                        if (userGroupRight.IsDelete)
+                        {
+                            var CountryToDelete = await _countryService.GetCountryByIdAsync(CompanyId, CountryId, UserId);
 
-                    if (CountryToDelete == null)
-                        return NotFound($"M_Country with Id = {CountryId} not found");
+                            if (CountryToDelete == null)
+                                return NotFound($"M_Country with Id = {CountryId} not found");
 
-                    var sqlResponce = await _countryService.DeleteCountryAsync(CompanyId, CountryToDelete, UserId);
-                    // Remove data from cache by key
-                    _memoryCache.Remove($"Country_{CountryId}");
-                    return Ok(sqlResponce);
+                            var sqlResponce = await _countryService.DeleteCountryAsync(CompanyId, CountryToDelete, UserId);
+                            // Remove data from cache by key
+                            _memoryCache.Remove($"Country_{CountryId}");
+                            return Ok(sqlResponce);
+                        }
+                        else
+                        {
+                            return NotFound("Users do not have a access to delete");
+                        }
+                    }
+                    else
+                    {
+                        return NotFound("Users not have a access for this screen");
+                    }
                 }
                 else
                 {
